Начинает заменят поинты на подель поитов
